pyproject.toml
[project]
name = "ft-agui"
version = "0.1.0"
description = "Easy AGUI integration for FastHTML applications - as simple as pydantic-ai's to_web()!"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Christoffer Björkskog", email = "christoffer.bjorkskog@novia.fi" }
]
keywords = ["fasthtml", "agui", "pydantic-ai", "chat", "websocket", "ui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: User Interfaces",
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-ai>=1.34.0",
    "python-fasthtml>=0.12.36",
    "ag-ui-core>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/Novia-RDI-Seafaring/ft-ag-ui"
Repository = "https://github.com/Novia-RDI-Seafaring/ft-ag-ui"
Documentation = "https://github.com/Novia-RDI-Seafaring/ft-ag-ui#readme"
"Bug Reports" = "https://github.com/Novia-RDI-Seafaring/ft-ag-ui/issues"


[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
